[Spark] Address review: dedup, dead-code, comment placement, file split
Addresses yili-db's review feedback on the view connector:

- Remove the meaningless `== 200` from `deleteTable` in both `UCProxy.dropTable`
  and `dropView`. `deleteTable` returns the (empty) response body, not an HTTP
  status; a real failure throws ApiException. Call it for its side effect and
  return true.
- Delete the empty `private object UCProxy` (three orphaned scaladoc blocks, no
  members; the methods they described live in the 4.2 `UCProxyViewSupport`).
- Move `UCColumnJson` out of `UCSingleCatalog.scala` into its own shared file
  `UCColumnJson.scala` (kept `private[spark]`; uses only public V2 APIs so it
  stays in the shared source tree, not a shim dir). Drop the now-unused json4s
  AST import from `UCSingleCatalog.scala`.
- Split the misplaced comment above `UCProxy.createTable`: keep the createTable
  rationale there, move the rename-stub note up to `renameTable`.
- De-duplicate `extractSqlConfigs`: hoist the byte-identical copy to the
  `UCSingleCatalog` companion (no Spark-version-specific types) and call it from
  the 4.2 view shim; remove both local copies.
- Remove the unused `isViewCommandsSupportedTableType(String)` overload (all call
  sites use the `TableType` overload; the create path gates via
  `sparkViewTypeToUcTableType`). Simplify its scaladoc.
- Document that `buildColumnInfos` intentionally mirrors `UCProxy.createTable`'s
  column loop but operates on the Spark-4.2-only V2 `Column`/`TableInfo` surface,
  so the near-duplication can't be merged without breaking 4.0/4.1.
- Cosmetics: fix the `ViewSupport` scaladoc bullet alignment and a double empty
  line in `UCSingleCatalog`.

Compiles on Spark 4.0.0 and 4.2.0-SNAPSHOT; UCProxySuite 18/19 on 4.2 (the one
failure, testLoadTableTriggersExactlyOneGetTableRpc, is a pre-existing
mock-URI/credential-vending env issue unrelated to these changes).
